# openapi.yaml
openapi: 3.0.1
info:
  title: Paperless Rest Server
  version: '1.0'
paths:
...
  /api/storage_paths:
    get:
      tags:
        - Config
      operationId: GetStoragePaths
      parameters:
        - name: page
          in: query
          style: form
          schema:
            type: integer
            format: int32
        - name: full_perms
          in: query
          style: form
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                  next:
                    type: boolean
                  previous:
                    type: boolean
                  all:
                    type: array
                    items:
                      type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        slug:
                          type: string
                        name:
                          type: string
                        path:
                          type: string
                        match:
                          type: string
                        matching_algorithm:
                          type: integer
                        is_insensitive:
                          type: boolean
                        document_count:
                          type: integer
                        owner:
                          type: integer
                        permissions:
                          type: object
                          properties:
                            view:
                              type: object
                              properties:
                                users:
                                  type: array
                                  items:
                                    type: integer
                                groups:
                                  type: array
                                  items: {}
                              required:
                                - users
                                - groups
                            change:
                              type: object
                              properties:
                                users:
                                  type: array
                                  items:
                                    type: integer
                                groups:
                                  type: array
                                  items: {}
                              required:
                                - users
                                - groups
                          required:
                            - view
                            - change
                      required:
                        - id
                        - slug
                        - name
                        - path
                        - match
                        - matching_algorithm
                        - is_insensitive
                        - document_count
                        - owner
                        - permissions
                required:
                  - count
                  - next
                  - previous
                  - all
                  - results
              examples:
                default:
                  value: >-
                    {"count":1,"next":null,"previous":null,"all":[1],"results":[{"id":1,"slug":"testdir","name":"testdir","path":"testdirp","match":"","matching_algorithm":6,"is_insensitive":true,"document_count":1,"owner":3,"permissions":{"view":{"users":[3],"groups":[]},"change":{"users":[3],"groups":[]}}}]}
    post:
      tags:
        - Config
      operationId: CreateStoragePath
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                path:
                  type: string
                matching_algorithm:
                  type: integer
                match:
                  type: string
                is_insensitive:
                  type: boolean
                owner:
                  type: integer
              required:
                - name
                - path
                - matching_algorithm
                - match
                - is_insensitive
                - owner
            examples:
              default:
                value: >-
                  {"name":"testdeldir","path":"testdeldir","matching_algorithm":6,"match":"","is_insensitive":true,"owner":3}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  slug:
                    type: string
                  name:
                    type: string
                  path:
                    type: string
                  match:
                    type: string
                  matching_algorithm:
                    type: integer
                  is_insensitive:
                    type: boolean
                  owner:
                    type: integer
                  user_can_change:
                    type: boolean
                required:
                  - id
                  - slug
                  - name
                  - path
                  - match
                  - matching_algorithm
                  - is_insensitive
                  - owner
                  - user_can_change
              examples:
                default:
                  value: >-
                    {"id":2,"slug":"testdeldir","name":"testdeldir","path":"testdeldir","match":"","matching_algorithm":6,"is_insensitive":true,"owner":3,"user_can_change":true}
  /api/storage_paths/{id}:
    put:
      tags:
        - Config
      operationId: UpdateStoragePath
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                slug:
                  type: string
                name:
                  type: string
                path:
                  type: string
                match:
                  type: string
                matching_algorithm:
                  type: integer
                is_insensitive:
                  type: boolean
                document_count:
                  type: integer
                owner:
                  type: integer
                permissions:
                  type: object
                  properties:
                    view:
                      type: object
                      properties:
                        users:
                          type: array
                          items:
                            type: integer
                        groups:
                          type: array
                          items: {}
                      required:
                        - users
                        - groups
                    change:
                      type: object
                      properties:
                        users:
                          type: array
                          items:
                            type: integer
                        groups:
                          type: array
                          items: {}
                      required:
                        - users
                        - groups
                  required:
                    - view
                    - change
                permissions_form:
                  type: object
                  properties:
                    owner:
                      type: integer
                  required:
                    - owner
              required:
                - id
                - slug
                - name
                - path
                - match
                - matching_algorithm
                - is_insensitive
                - document_count
                - owner
                - permissions
                - permissions_form
            examples:
              default:
                value: >-
                  {"id":1,"slug":"testdir","name":"testdir2","path":"testdirp","match":"","matching_algorithm":6,"is_insensitive":true,"document_count":1,"owner":3,"permissions":{"view":{"users":[3],"groups":[]},"change":{"users":[3],"groups":[]}},"permissions_form":{"owner":3,}}
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  slug:
                    type: string
                  name:
                    type: string
                  path:
                    type: string
                  match:
                    type: string
                  matching_algorithm:
                    type: integer
                  is_insensitive:
                    type: boolean
                  document_count:
                    type: integer
                  owner:
                    type: integer
                  user_can_change:
                    type: boolean
                required:
                  - id
                  - slug
                  - name
                  - path
                  - match
                  - matching_algorithm
                  - is_insensitive
                  - document_count
                  - owner
                  - user_can_change
              examples:
                default:
                  value: >-
                    {"id":1,"slug":"testdir2","name":"testdir2","path":"testdirp","match":"","matching_algorithm":6,"is_insensitive":true,"document_count":1,"owner":3,"user_can_change":true}
    delete:
      tags:
        - Config
      operationId: DeleteStoragePath
      parameters:
        - name: id
          in: path
          required: true
          style: simple
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: Success
...